The explanation under the transcript at 60:27:59
Shouldn't it be "But this has the potential to alter the trajectory of the spacecraft, and this far to the Moon very small adjustments can turn into miles of difference at the very far point" instead of "But this has the potential to alter the trajectory of the spacecraft, and this close to the Moon very small adjustments can turn into miles of difference at the very far point". Changed close to far.
The closer you are from the pericynthion, the more DeltaV is required to alter the orbit, subsequently farther means less DeltaV required so more chances to bring the pericynthion under 60 with small manuevers.